move CellType up to eval namespace
