+++ eval/src/vespa/eval/eval/cell_type.cpp
@@ -1,0 +1,3 @@
+// Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#include "cell_type.h"
